requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.11.7
pillow==10.4.0
opencv-python==4.12.0.88
pydicom==2.4.3
openai>=1.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
numpy==2.2.0
PyYAML==6.0.2